Added freeform question functionality
Freeform questions can be evoked by mentioning the bot with the command 'freeform'. An Adaptive Card will be returned to the chat with a text input box for a question to submit to the Space. On sending, everyone in the Space will see an Adaptive Card with the freeform question and a text input box to submit their answer to the question. The user that evoked the question receives a DM inviting them to view the responses to their question by clicking the button on the card. The user can see the responses whenever they'd like to see them.