Poll Rework: Completed poll data retrieval
new function getPollResultsCard will retrieve the results and create a new object array of each choice, containing the count of that choice, and a string array of the names of those who chose them. Also, the object array contains 'otherAnswerArray', which holds the names and specified answers for the people who chose the option 'Other' in the poll.

Still to complete for Poll rework: Displaying the results on a pollRequest