Card upgrades, bug fixes
Highfivecard and Birthdaycard now work on mentions instead of emails, which should be way less of a hassle to invoke.

Fixed an issue where poll results would display the 'Other' option despite not having 'Other' in the poll

Fixed an issue where not including anyone in the highfivecard or birthdaycard commands would crash the bot.